+++ webpack/webpack.config.js
@@ -12,10 +12,17 @@ let gitHash = childProcess.execSync(gitCmd).toString().substring(0, 7);
 
 const staticPath = path.resolve(__dirname, "../src/senaite/databox/browser/static");
 
-const devMode = process.env.NODE_ENV !== 'production';
+const devMode = process.env.mode == "development";
+const prodMode = process.env.mode == "production";
+const mode = process.env.mode;
+console.log(`RUNNING WEBPACK IN '${mode}' MODE`);
 
 
 module.exports = {
+  // https://webpack.js.org/configuration/devtool
+  devtool: devMode ? "eval" : "",
+  // https://webpack.js.org/configuration/mode/#usage
+  mode: mode,
   context: path.resolve(__dirname, "app"),
   entry: {
     "senaite.databox": [
@@ -24,8 +31,7 @@ module.exports = {
     ],
   },
   output: {
-    filename: `[name]-${gitHash}.js`,
-    filename: gitHash ? `[name]-${gitHash}.js` : "[name].js",
+    filename: "[name].js",
     path: path.resolve(staticPath, "bundles"),
     publicPath: "/++plone++senaite.databox.static/bundles"
   },
@@ -53,10 +59,7 @@ module.exports = {
         use: [
           {
             // https://webpack.js.org/plugins/mini-css-extract-plugin/
-            loader: MiniCssExtractPlugin.loader,
-            options: {
-              hmr: process.env.NODE_ENV === "development"
-            },
+            loader: MiniCssExtractPlugin.loader
           },
           {
             // https://webpack.js.org/loaders/css-loader/
@@ -95,8 +98,7 @@ module.exports = {
     }),
     // https://webpack.js.org/plugins/mini-css-extract-plugin/
     new MiniCssExtractPlugin({
-      filename: devMode ? "[name].css" : "[name].[hash].css",
-      chunkFilename: devMode ? "[id].css" : "[id].[hash].css",
+      filename: "[name].css"
     }),
     // https://webpack.js.org/plugins/copy-webpack-plugin/
     // new CopyPlugin({