+++ docs/Changelog.rst
@@ -1,6 +1,7 @@
 1.1.0 (unreleased)
 ------------------
 
+- #19 Fix columns and active-tab handling
 - #17 Fix selection of non-sortable indexes
 - #16 Added upgrade step handler
 - #14 Reduce conflict errors when creating temporary objects